+++ src/jtl/Connector/Presta/Controller/CustomerOrder.php
@@ -11,8 +11,8 @@ class CustomerOrder extends BaseController
 			FROM '._DB_PREFIX_.'orders o
 			LEFT JOIN '._DB_PREFIX_.'currency c ON c.id_currency = o.id_currency
 			LEFT JOIN '._DB_PREFIX_.'carrier s ON s.id_carrier = o.id_carrier
-			LEFT JOIN jtl_connector_link_customer_order l ON o.id_order = l.endpointId AND l.type = 4
-            WHERE l.hostId IS NULL';
+			LEFT JOIN jtl_connector_link_customer_order l ON o.id_order = l.endpoint_id
+            WHERE l.host_id IS NULL';
 
         if (!empty(\Configuration::get('jtlconnector_from_date'))) {
             $query .= ' && o.date_add >= "'.\Configuration::get('jtlconnector_from_date').'"';
@@ -39,8 +39,8 @@ class CustomerOrder extends BaseController
     {
         $query = 'SELECT COUNT(*)
 			FROM '._DB_PREFIX_.'orders o
-			LEFT JOIN jtl_connector_link_customer_order l ON o.id_order = l.endpointId AND l.type = 4
-            WHERE l.hostId IS NULL';
+			LEFT JOIN jtl_connector_link_customer_order l ON o.id_order = l.endpoint_id
+            WHERE l.host_id IS NULL';
 
         if (!empty(\Configuration::get('jtlconnector_from_date'))) {
             $query .= ' && o.date_add >= "'.\Configuration::get('jtlconnector_from_date').'"';