+++ src/jtl/Connector/Presta/Controller/Payment.php
@@ -9,8 +9,8 @@ class Payment extends BaseController
 			SELECT p.*, o.id_order
 			FROM '._DB_PREFIX_.'order_payment p
 			LEFT JOIN '._DB_PREFIX_.'orders o ON o.reference = p.order_reference
-			LEFT JOIN jtl_connector_link_payment l ON p.id_order_payment = l.endpointId AND l.type = 512
-            WHERE l.hostId IS NULL AND p.transaction_id != ""
+			LEFT JOIN jtl_connector_link_payment l ON p.id_order_payment = l.endpoint_id
+            WHERE l.host_id IS NULL AND p.transaction_id != ""
             LIMIT '.$limit
         );
 
@@ -30,8 +30,8 @@ class Payment extends BaseController
         return $this->db->getValue('
 			SELECT COUNT(*)
 			FROM '._DB_PREFIX_.'order_payment p
-			LEFT JOIN jtl_connector_link_payment l ON p.id_order_payment = l.endpointId AND l.type = 512
-            WHERE l.hostId IS NULL AND p.transaction_id != ""
+			LEFT JOIN jtl_connector_link_payment l ON p.id_order_payment = l.endpoint_id
+            WHERE l.host_id IS NULL AND p.transaction_id != ""
         ');
     }
 }