+++ src/jtl/Connector/Presta/Controller/Specific.php
@@ -21,8 +21,8 @@ class Specific extends BaseController
         $specificsIds = $this->db->executeS(sprintf('
 			SELECT v.id_feature
 			FROM %sfeature_value v
-			LEFT JOIN jtl_connector_link_product l ON v.id_feature = l.endpointId AND l.type = 128
-            WHERE l.hostId IS NULL AND v.custom = 0
+			LEFT JOIN jtl_connector_link_product l ON v.id_feature = l.endpoint_id
+            WHERE l.host_id IS NULL AND v.custom = 0
             GROUP BY v.id_feature
             LIMIT %s',
             _DB_PREFIX_,
@@ -270,8 +270,8 @@ class Specific extends BaseController
         SELECT COUNT(*)
         FROM (SELECT v.id_feature
               FROM %sfeature_value v
-              LEFT JOIN jtl_connector_link_specific l ON v.id_feature = l.endpointId AND l.type = 128
-              WHERE l.hostId IS NULL AND v.custom = 0
+              LEFT JOIN jtl_connector_link_specific l ON v.id_feature = l.endpoint_id
+              WHERE l.host_id IS NULL AND v.custom = 0
               GROUP BY v.id_feature) as Z',
             _DB_PREFIX_
         ));